src/streamlit_app/requirements.txt
streamlit==1.39.0
pandas==1.5.3
numpy==1.24.1
requests==2.28.2
python-jose
altair
